refactor: MVC 패턴으로 변경
